Fix problem with syncronous body streaming
