+++ tremor-connectors/Cargo.toml
@@ -20,7 +20,7 @@ tremor-system = { path = "../tremor-system", version = "0.13.0-rc.24" }
 tokio = { version = "1.34", default-features = false }
 beef = { version = "0.5", default-features = false }
 value-trait = { version = "0.8", default-features = false }
-futures = { version = "0.3", default-features = false }
+futures = { version = "0.3.30", default-features = false, features = ["std"] }
 halfbrown = { version = "0.2", default-features = false }
 simd-json = { version = "0.13", default-features = false }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
@@ -95,6 +95,7 @@ reqwest = { version = "0.11", optional = true, default-features = false, feature
 ] }
 base64 = { version = "0.22", optional = true, default-features = false }
 dashmap = { version = "5.5", optional = true, default-features = false }
+async-stream = { version = "0.3", optional = true, default-features = false }
 
 # kafka
 rdkafka = { version = "0.36", optional = true, features = [
@@ -167,7 +168,7 @@ bytes = "1.0"
 tempfile = { version = "3.8", default-features = false }
 env_logger = "0.11"
 tremor-connectors-test-helpers = { path = "../tremor-connectors-test-helpers", version = "0.13.0-rc.24" }
-tide = { version = "0.16", default-features = false }      # TODO remove tide from TestHttpServer
+tide = { version = "0.16", default-features = false } # TODO remove tide from TestHttpServer
 tokio = { version = "1.34", default-features = false, features = [
     "full",
     "test-util",
@@ -231,6 +232,7 @@ http = [
     "dep:hyper",
     "dep:mime",
     "dep:reqwest",
+    "dep:async-stream",
     "hyper/http1",
     "hyper/http2",
     "tls",
