+++ tremor-connectors/src/impls/http/meta.rs
@@ -12,53 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    impls::http::{client, utils::Body, utils::RequestId},
-    sink::EventSerializer,
-    utils::mime::MimeCodecMap,
-};
-use either::Either;
-use http::{
-    header::{self, HeaderName, ToStrError},
-    HeaderMap, Uri,
-};
-use http_body_util::BodyStream;
-use hyper::{header::HeaderValue, Method, Request, Response};
+use crate::utils::mime::MimeCodecMap;
+use http::{header::ToStrError, HeaderMap};
+use hyper::{header::HeaderValue, Request, Response};
 use mime::Mime;
-use tokio::sync::mpsc::{channel, Receiver, Sender};
 use tremor_config::NameWithConfig;
-use tremor_system::qsize;
 use tremor_value::prelude::*;
-
-use super::utils::body_from_bytes;
-
-/// HTTP Connector Errors
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    /// Invalid HTTP Method
-    #[error("Invalid HTTP Method")]
-    InvalidMethod,
-    /// Invalid HTTP URL
-    #[error("Invalid HTTP URL")]
-    InvalidUrl,
-    /// Request already consumed
-    #[error("Request already consumed")]
-    RequestAlreadyConsumed,
-    /// Response already consumed
-    #[error("Response already consumed")]
-    ResponseAlreadyConsumed,
-}
-
-/// Utility for building an HTTP request from a possibly batched event
-/// and some configuration values
-pub(crate) struct HttpRequestBuilder {
-    request_id: RequestId,
-    builder: Option<http::request::Builder>,
-    //request: Option<hyper::Request<BodyStream<Body>>>,
-    chunk_tx: Sender<Vec<u8>>,
-    chunk_rx: Receiver<Vec<u8>>,
-    codec_overwrite: Option<NameWithConfig>,
-}
 
 #[derive(Clone)]
 pub(crate) struct HeaderValueValue<'v> {
@@ -110,136 +69,6 @@ impl<'v> Iterator for HeaderValueValue<'v> {
     }
 }
 
-// TODO: do some deduplication with SinkResponse
-impl HttpRequestBuilder {
-    pub(super) fn new(
-        request_id: RequestId,
-        meta: Option<&Value>,
-        codec_map: &MimeCodecMap,
-        config: &client::Config,
-    ) -> anyhow::Result<Self> {
-        let request_meta = meta.get("request");
-        let method = if let Some(method_v) = request_meta.get("method") {
-            Method::from_bytes(method_v.as_bytes().ok_or(Error::InvalidMethod)?)?
-        } else {
-            config.method.0.clone()
-        };
-        let uri: Uri = if let Some(url_v) = request_meta.get("url") {
-            url_v.as_str().ok_or(Error::InvalidUrl)?.parse()?
-        } else {
-            config.url.to_string().parse()?
-        };
-        let mut request = Request::builder().method(method).uri(uri);
-
-        // first insert config headers
-        for (config_header_name, config_header_values) in &config.headers {
-            match &config_header_values.0 {
-                Either::Left(config_header_values) => {
-                    for header_value in config_header_values {
-                        request =
-                            request.header(config_header_name.as_str(), header_value.as_str());
-                    }
-                }
-                Either::Right(header_value) => {
-                    request = request.header(config_header_name.as_str(), header_value.as_str());
-                }
-            }
-        }
-        let headers = request_meta.get("headers");
-
-        // build headers
-        if let Some(headers) = headers.as_object() {
-            for (name, values) in headers {
-                let name = HeaderName::from_bytes(name.as_bytes())?;
-                for value in HeaderValueValue::new(values) {
-                    request = request.header(&name, value);
-                }
-            }
-        }
-
-        let header_content_type = content_type(request.headers_ref())?;
-
-        let (codec_overwrite, content_type) =
-            consolidate_mime(header_content_type.clone(), codec_map);
-
-        // set the content type if it is not set yet
-        if header_content_type.is_none() {
-            if let Some(ct) = content_type {
-                request = request.header(header::CONTENT_TYPE, ct.to_string());
-            }
-        }
-        // handle AUTH
-        if let Some(auth_header) = config.auth.as_header_value()? {
-            request = request.header(hyper::header::AUTHORIZATION, auth_header);
-        }
-
-        let (chunk_tx, chunk_rx) = channel::<Vec<u8>>(qsize());
-
-        // extract headers
-        // determine content-type, override codec and chunked encoding
-        Ok(Self {
-            request_id,
-            builder: Some(request),
-            chunk_tx,
-            chunk_rx,
-            codec_overwrite,
-        })
-    }
-
-    pub(super) async fn append<'event>(
-        &mut self,
-        value: &'event Value<'event>,
-        meta: &'event Value<'event>,
-        ingest_ns: u64,
-        serializer: &mut EventSerializer,
-    ) -> anyhow::Result<()> {
-        let chunks = serializer
-            .serialize_for_stream_with_codec(
-                value,
-                meta,
-                ingest_ns,
-                self.request_id.get(),
-                self.codec_overwrite.as_ref(),
-            )
-            .await?;
-        self.append_data(chunks).await
-    }
-
-    async fn append_data(&mut self, chunks: Vec<Vec<u8>>) -> anyhow::Result<()> {
-        for chunk in chunks {
-            self.chunk_tx.send(chunk).await?;
-        }
-        Ok(())
-    }
-
-    pub(super) fn take_request(&mut self) -> Result<Request<BodyStream<Body>>, Error> {
-        let mut data = Vec::new();
-
-        // drain the channel
-        while let Ok(chunk) = self.chunk_rx.try_recv() {
-            data.extend(chunk);
-        }
-
-        let builder = self.builder.take().ok_or(Error::RequestAlreadyConsumed)?;
-        let body_stream = BodyStream::new(body_from_bytes(data));
-        let request = builder
-            .body(body_stream)
-            .map_err(|_| Error::RequestAlreadyConsumed)?;
-        Ok(request)
-    }
-    /// Finalize and send the response.
-    /// In the chunked case we have already sent it before.
-    ///
-    /// After calling this function this instance shouldn't be used anymore
-    pub(super) async fn finalize(
-        &mut self,
-        serializer: &mut EventSerializer,
-    ) -> anyhow::Result<()> {
-        // finalize the stream
-        let rest = serializer.finish_stream(self.request_id.get())?;
-        self.append_data(rest).await
-    }
-}
 /// Extract the content type from the headers
 /// # Errors
 ///   if the content type is not a valid string
@@ -330,40 +159,4 @@ pub(super) fn extract_response_meta<B>(
     meta.try_insert("headers", headers);
     meta.try_insert("version", format!("{:?}", response.version()));
     Ok(meta)
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::CodecReq;
-    use crate::ConnectorType;
-    use tremor_common::alias;
-    use tremor_config::Impl;
-
-    use super::*;
-    #[tokio::test(flavor = "multi_thread")]
-    async fn builder() -> anyhow::Result<()> {
-        let request_id = RequestId::new(42);
-        let meta = None;
-        let codec_map = MimeCodecMap::default();
-        let c = literal!({"headers": {
-            "cake": ["black forst", "cheese"],
-            "pie": "key lime"
-        }});
-        let mut s = EventSerializer::new(
-            None,
-            CodecReq::Optional("json"),
-            vec![],
-            &ConnectorType("http".into()),
-            &alias::Connector::new("flow", "http"),
-        )?;
-        let config = client::Config::new(&c)?;
-
-        let mut b = HttpRequestBuilder::new(request_id, meta, &codec_map, &config)?;
-
-        let r = b.take_request()?;
-        b.finalize(&mut s).await?;
-        assert_eq!(r.headers().get_all("pie").iter().count(), 1);
-        assert_eq!(r.headers().get_all("cake").iter().count(), 2);
-        Ok(())
-    }
 }
